Fixed warnings

+++ Planning/RobotTimeScaling.cpp
@@ -963,7 +963,7 @@ void EvaluateMultiPath(Robot& robot,const MultiPath& path,Real t,Config& q,Real 
   Real duration,param;
   int seg=path.Evaluate(t,curve,duration,param,MultiPath::InterpLinear);
   if(seg < 0) seg = 0;
-  if(seg >= path.sections.size()) seg = (int)path.sections.size()-1;
+  if(seg >= (int)path.sections.size()) seg = (int)path.sections.size()-1;
   curve.Eval(param,q);
 
   //solve for constraints
